pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.0.0",
    "fastapi[standard]>=0.128.4",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]
